refactor(executor): extract execute_type_safe method to dedicated module
Moved the execute_type_safe method from the main executor.rs file to
a new executor/type_safe_execution.rs module. This is part of the
larger effort to break down the 3000+ line executor.rs file into
smaller, more focused modules.

The refactoring maintains the exact same functionality and API - only
the location of the code has changed. All tests continue to pass
without modification.

Part of #53
